requirements-dev.txt
# Development tools
black
jupyter
mypy
pydocstyle
pylint
pytest
pytest-cov
pytest-html

# Packaging requirements
setuptools
twine

# Sphinx requirements
numpydoc
sphinx
sphinx-rtd-theme
